move ripgrep to own file
