+++ app/src/remote_server/mod.rs
@@ -26,6 +26,8 @@ pub mod git_status_proto;
 #[cfg(not(target_family = "wasm"))]
 pub(crate) mod handoff_snapshot;
 #[cfg(not(target_family = "wasm"))]
+mod ripgrep_search;
+#[cfg(not(target_family = "wasm"))]
 pub mod server_buffer_tracker;
 #[cfg(not(target_family = "wasm"))]
 pub mod server_model;
